refactor(utils): utils 파일을 lib 폴더로 이동
